VideoCutterApp/requirements.txt
PySide6==6.8.0
python-vlc==3.0.20123
psutil==6.0.0